+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockKeycardReader.java
@@ -5,6 +5,7 @@ import java.util.Random;
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
+import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.gui.GuiHandler;
 import net.geforcemods.securitycraft.items.ItemKeycard;
 import net.geforcemods.securitycraft.tileentity.TileEntityKeycardReader;
@@ -127,9 +128,10 @@ public class BlockKeycardReader extends BlockDisguisable {
 
 	public ITextComponent insertCard(World world, BlockPos pos, IBlockState state, TileEntityKeycardReader te, ItemStack stack, EntityPlayer player) {
 		NBTTagCompound tag = stack.getTagCompound();
+		Owner keycardOwner = new Owner(tag.getString("ownerName"), tag.getString("ownerUUID"));
 
 		//owner of this keycard reader and the keycard reader the keycard got linked to do not match
-		if ((ConfigHandler.enableTeamOwnership && !PlayerUtils.areOnSameTeam(te.getOwner().getName(), tag.getString("ownerName"))) || !te.getOwner().getUUID().equals(tag.getString("ownerUUID")))
+		if ((ConfigHandler.enableTeamOwnership && !PlayerUtils.areOnSameTeam(te.getOwner(), keycardOwner)) || !te.getOwner().getUUID().equals(keycardOwner.getUUID()))
 			return new TextComponentTranslation("messages.securitycraft:keycardReader.differentOwner");
 
 		//the keycard's signature does not match this keycard reader's