+++ src/main/java/net/geforcemods/securitycraft/items/ItemUniversalKeyChanger.java
@@ -37,7 +37,7 @@ public class ItemUniversalKeyChanger extends Item {
 				return EnumActionResult.SUCCESS;
 			}
 			else if (!(te.getBlockType() instanceof BlockDisguisable) || (((ItemBlock) ((BlockDisguisable) te.getBlockType()).getDisguisedStack(world, pos).getItem()).getBlock() instanceof BlockDisguisable)) {
-				PlayerUtils.sendMessageToPlayer(player, Utils.localize("item.securitycraft:universalKeyChanger.name"), Utils.localize("messages.securitycraft:notOwned", PlayerUtils.getOwnerComponent(((IOwnable) world.getTileEntity(pos)).getOwner().getName())), TextFormatting.RED);
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize("item.securitycraft:universalKeyChanger.name"), Utils.localize("messages.securitycraft:notOwned", PlayerUtils.getOwnerComponent(((IOwnable) te).getOwner())), TextFormatting.RED);
 				return EnumActionResult.SUCCESS;
 			}
 		}