+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityRetinalScanner.java
@@ -74,7 +74,7 @@ public class TileEntityRetinalScanner extends TileEntityDisguisable implements I
 						viewingPlayer = PlayerUtils.getSkullOwner(player);
 
 					if (!isOwnedBy(viewingPlayer) && !isAllowed(viewingPlayer.getName())) {
-						PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.retinalScanner), Utils.localize("messages.securitycraft:retinalScanner.notOwner", PlayerUtils.getOwnerComponent(getOwner().getName())), TextFormatting.RED);
+						PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.retinalScanner), Utils.localize("messages.securitycraft:retinalScanner.notOwner", PlayerUtils.getOwnerComponent(getOwner())), TextFormatting.RED);
 						return true;
 					}