+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityScannerDoor.java
@@ -48,7 +48,7 @@ public class TileEntityScannerDoor extends TileEntitySpecialDoor implements IVie
 					viewingPlayer = PlayerUtils.getSkullOwner(player);
 
 				if (!isOwnedBy(viewingPlayer) && !isAllowed(viewingPlayer.getName())) {
-					PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.retinalScanner), Utils.localize("messages.securitycraft:retinalScanner.notOwner", PlayerUtils.getOwnerComponent(getOwner().getName())), TextFormatting.RED);
+					PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.retinalScanner), Utils.localize("messages.securitycraft:retinalScanner.notOwner", PlayerUtils.getOwnerComponent(getOwner())), TextFormatting.RED);
 					return true;
 				}