+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.6 of SecurityCraft--------------------------
 
+- New: Support for teams from FTB Utilities when having team ownership enabled in the config
 - Fix: Turning off camera rotation leads to extreme stuttering of the camera
 - Fix: Shift-clicking a potion in the brewing stand screen puts the potion into the wrong slot